restore api

+++ index.js
@@ -53,7 +53,7 @@ app.get('/data/:token?', async (req, res) => {
             });
 
             await pool.query(`SELECT videos.id, videos.employment as name, videos.description, videos.publish_date, videos.end_date, videos.is_active, videos.confirm,
-                (SELECT COUNT(id) FROM likes WHERE video_id = videos.id ) as likes 
+                (SELECT COUNT(id) FROM likes WHERE video_id = videos.id )::int as likes 
                 FROM videos INNER JOIN users ON videos.user_id = users.id
                 WHERE users.token = '${req.params.token}'`)
             .then(videos =>{
@@ -73,7 +73,7 @@ app.get('/data/:token?', async (req, res) => {
         });
         
         await pool.query(`SELECT videos.id, videos.index, videos.user_id, users.username, users.email, users.employment, videos.employment as name, users.token, videos.types, 
-            (SELECT COUNT(id) FROM likes WHERE video_id = videos.id ) as likes 
+            (SELECT COUNT(id) FROM likes WHERE video_id = videos.id )::int as likes 
             FROM videos INNER JOIN users ON videos.user_id = users.id
             WHERE videos.index = 0
             AND videos.is_active=TRUE 
@@ -83,7 +83,7 @@ app.get('/data/:token?', async (req, res) => {
         });
 
         await pool.query(`SELECT videos.id, videos.index, videos.user_id, users.username, users.email, users.employment, videos.employment as name, users.token, videos.description, videos.publish_date, videos.end_date, videos.types, 
-            (SELECT COUNT(id) FROM likes WHERE video_id = videos.id ) as likes
+            (SELECT COUNT(id) FROM likes WHERE video_id = videos.id )::int as likes
             FROM videos INNER JOIN users ON videos.user_id = users.id
             WHERE videos.index = 1
             AND videos.is_active=TRUE 
@@ -93,7 +93,7 @@ app.get('/data/:token?', async (req, res) => {
         });
 
         await pool.query(`SELECT videos.id, videos.index, videos.user_id, users.username, users.email, users.employment, videos.employment as name, users.token, videos.types,
-            (SELECT COUNT(id) FROM likes WHERE video_id = videos.id ) as likes
+            (SELECT COUNT(id) FROM likes WHERE video_id = videos.id )::int as likes
             FROM videos INNER JOIN users ON videos.user_id = users.id
             WHERE videos.index = 2
             AND videos.is_active=TRUE 
@@ -103,7 +103,7 @@ app.get('/data/:token?', async (req, res) => {
         });
 
         await pool.query(`SELECT videos.id, videos.index, videos.user_id, users.username, users.email, users.employment, videos.employment as name, users.token, videos.types,
-            (SELECT COUNT(id) FROM likes WHERE video_id = videos.id ) as likes
+            (SELECT COUNT(id) FROM likes WHERE video_id = videos.id )::int as likes
             FROM videos INNER JOIN users ON videos.user_id = users.id
             WHERE videos.index = 3
             AND videos.is_active=TRUE
